llm-module/requirements.txt
psycopg2-binary
spacy
requests
pika
google-generativeai
prometheus_client
sqlalchemy
pandas
redis